Update web pages' style and apply scrollbar
Created a web page template. Applied uniform style to web pages covering
various CSS topics. Applied a scrollbar to the div within the sidebar to
account for overflow and to allow users to access the "return to CSS
lessons" button. Adjusted the sidebar's style, changing the margins,
padding, and line height to make the in-page links readable.